Implemented constraint propagation
To integrate constraint propagation, I have been developing an unreleased sister library named pyDuck that implements a CSP solver with constraint propagation. Until I polish the code for an independent release, this will be the home for that library as well. The actual backtracking search is yet to be integrated into main system but it is implemented well in pyDuck.