Added ">" and truly undirectional substitutions
All goals were substituted with whatever was available at the moment of their addition to a given SystemEquations but now they are substituted with the substitutions given at the time of propagation or potentially search.

Also added "Greater than" operator to be used along with "+" and "==", more shall be added soon.